requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
distro==1.9.0
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
keyboard==0.13.5
MouseInfo==0.1.3
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.8.2
pydantic_core==2.20.1
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.2.0
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
python-dotenv==1.0.0
playsound3==3.2.3